Sessions controller created
Was created the sessions controller, this controller is necessary to
logout a user. When a user signs up, a session will be created and will
be required for the other views, the same when a user logs out. This
means, if the user is not logged in, then will be redirected to login
view.

When a user wants create an account, the user controller creates the
user and the session, then redirecto to root_path. The require_login
method that is present in the application controller executes this
method before any action and if the session exist, the nothing is
happen, but if not, then the user will be redirected to login path.

When a user wants to login, the sessions controller takes this action
and validates if the user with that credentials exist, if true, then
creates a new session and redirecto to root_path.

When a user wants to logout, the sessions controller takes this action
with the destroy method, just establish the session variable to nil and
then will be redirected to login path.